Fix labels data type

+++ model_api/cpp/models/src/detection_model_ssd.cpp
@@ -125,7 +125,7 @@ std::unique_ptr<ResultBase> ModelSSD::postprocessSingleOutput(InferenceResult& i
 std::unique_ptr<ResultBase> ModelSSD::postprocessMultipleOutputs(InferenceResult& infResult) {
     const float* boxes = infResult.outputsData[outputNames[0]].data<float>();
     size_t detectionsNum = infResult.outputsData[outputNames[0]].get_shape()[detectionsNumId];
-    const float* labels = infResult.outputsData[outputNames[1]].data<float>();
+    const int64_t* labels = infResult.outputsData[outputNames[1]].data<int64_t>();
     const float* scores = outputNames.size() > 2 ? infResult.outputsData[outputNames[2]].data<float>() : nullptr;
 
     DetectionResult* result = new DetectionResult(infResult.frameId, infResult.metaData);
@@ -158,7 +158,7 @@ std::unique_ptr<ResultBase> ModelSSD::postprocessMultipleOutputs(InferenceResult
             DetectedObject desc;
 
             desc.confidence = confidence;
-            desc.labelID = static_cast<int>(labels[i]);
+            desc.labelID = labels[i];
             desc.label = getLabelName(desc.labelID);
             desc.x = clamp(
                 round((boxes[i * objectSize] * widthScale - padLeft) * invertedScaleX),
